requirements.txt
Flask==2.0.2

Flask-SQLAlchemy==2.5.1

requests==2.26.0

beautifulsoup4==4.10.0

huggingface-hub-0.1.2

pyjwt-2.3.0